requirements.txt
selenium==4.19.0
pytest==8.1.1
allure-pytest==2.13.2
pyyaml==6.0.1
openpyxl==3.1.2
pytest-rerunfailures
